+++ crate_anon/preprocess/preprocess_pcmis.py
@@ -569,7 +569,7 @@ def get_pcmis_dd_settings(ddhint: DDHint) -> str:
 
 {sk.DDGEN_FORCE_LOWER_CASE} = False
 {sk.DDGEN_CONVERT_ODD_CHARS_TO_UNDERSCORE} = True
-"""  # noqa
+"""  # noqa: E501
 
 
 # =============================================================================