+++ docs/source/changelog.rst
@@ -1692,6 +1692,11 @@ Changes
   characters in the source database. Convert any unusual characters to ASCII in
   the destination database.
 
+- Update NLP handler to cope with remote NLPRP servers providing tabular_schema
+  data, and create local tables based on this, if desired. Change default
+  NLPRP server behaviour to use the more explicit format (object/dictionaries
+  by table, not plain arrays/lists of records).
+
 
 To do
 -----